backend/requirements.txt
alembic==1.18.4
annotated-doc==0.0.4
mangum>=0.17.0
annotated-types==0.7.0
anyio==4.13.0
bcrypt==5.0.0
boto3==1.42.78
botocore==1.42.78
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
ecdsa==0.19.2
fastapi==0.135.2
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
jmespath==1.1.0
Mako==1.3.10
MarkupSafe==3.0.3
passlib==1.7.4
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg2-binary==2.9.11
pyasn1==0.6.3
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.12.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose==3.5.0
python-multipart==0.0.22
rsa==4.9.1
s3transfer==0.16.0
six==1.17.0
SQLAlchemy==2.0.48
sqlmodel==0.0.37
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.42.0
